Enum encryptfile::RngMode

pub enum RngMode {
    Os,
    OsIssac,
    OsRandIssac,
    Func(Rc<Box<Fn() -> u8>>),
}

Controls how random numbers are generated whenever they are needed by this library. Currently this is only required when generating an initialization vector (InitializationVector::GenerateFromRng). Note, when decrypting, you do not need to specify this.

Variants

Os

Use the Os RNG only

OsIssac

Use a combination of the Os and Isaac64 generators. Isaac is seeded with the Os RNG, and the two RNGs are used to generate the resulting IV 50/50.

OsRandIssac

Use a combination of the Os, rand::random, and Isaac. Isaac is seeded with the Os Rng and rand. Os and Isaac are used to generate the resulting IV 50/50.

Func

Use the specified function to generate random u8 values. The function should return a random u8 each time it is called.